agents/interview-agent/pyproject.toml
[project]
name = "uv-platform-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioboto3>=14.3.0",
    "aiohttp>=3.12.13",
    "livekit-agents>=1.1.1",
    "livekit-plugins-deepgram>=1.1.1",
    "livekit-plugins-noise-cancellation>=0.2.4",
    "livekit-plugins-openai>=1.1.1",
    "livekit-plugins-silero>=1.1.1",
    "livekit-plugins-turn-detector>=1.1.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
]
